Restore listeners, timeout after 30s on send
The use of STM and forkIO throughout time-warp makes it very possible
and plausible that data becomes stuck in chans as threads are blocked
indefinitely on those chans (one of the threads which feeds or drains it
has stopped doing so). If this happens, we probably wouldn't see it in
the logs, and a blocked indefinitely exception would not be raised,
because it's likely that the 'ThreadId' of the blocked thread is
retained somwhere, whether by 'seqConcurrentlyK' in 'sendToNeighbors' or
by the "interrupter" of some thread job.