[CSL-811] Generalize 'verifyHeader' block header test
In order to generate a block header and associated 'VerifyHeaderParams'
for the 'verifyHeader' function, the previous 'Arbitrary HeaderAndParams'
instance simply fetched the first three blocks of a valid generated
headerchain. To make it more general, it now randomly chooses a header
in the chain instead of always going for the beginning.